# openapi-health-score.json
{"openapi":"3.1.0","info":{"title":"RepoIQ API","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}},"/auth/register":{"post":{"tags":["authentication"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/profile":{"get":{"tags":["authentication"],"summary":"Profile","operationId":"profile_auth_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/repositories/import":{"post":{"tags":["repositories"],"summary":"Import Repository","operationId":"import_repository_repositories_import_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepositoryImportRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepositoryRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/repositories/{repository_id}":{"get":{"tags":["repositories"],"summary":"Get Repository Details","operationId":"get_repository_details_repositories__repository_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"repository_id","in":"path","required":true,"schema":{"type":"integer","title":"Repository Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RepositoryRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/repositories/{repository_id}/health-score":{"post":{"tags":["repositories"],"summary":"Generate Repository Health Score","operationId":"generate_repository_health_score_repositories__repository_id__health_score_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"repository_id","in":"path","required":true,"schema":{"type":"integer","title":"Repository Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthScoreResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthScoreCheck":{"properties":{"name":{"type":"string","title":"Name"},"passed":{"type":"boolean","title":"Passed"},"weight":{"type":"integer","title":"Weight"},"detail":{"type":"string","title":"Detail"}},"type":"object","required":["name","passed","weight","detail"],"title":"HealthScoreCheck"},"HealthScoreResponse":{"properties":{"id":{"type":"integer","title":"Id"},"score":{"type":"number","title":"Score"},"breakdown":{"items":{"$ref":"#/components/schemas/HealthScoreCheck"},"type":"array","title":"Breakdown"}},"type":"object","required":["id","score","breakdown"],"title":"HealthScoreResponse"},"LoginRequest":{"properties":{"email":{"type":"string","maxLength":255,"minLength":3,"title":"Email"},"password":{"type":"string","maxLength":72,"minLength":1,"title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"RegistrationRequest":{"properties":{"username":{"type":"string","maxLength":100,"minLength":3,"title":"Username"},"email":{"type":"string","maxLength":255,"minLength":3,"title":"Email"},"password":{"type":"string","maxLength":72,"minLength":8,"title":"Password"}},"type":"object","required":["username","email","password"],"title":"RegistrationRequest"},"RepositoryImportRequest":{"properties":{"github_url":{"type":"string","title":"Github Url"}},"type":"object","required":["github_url"],"title":"RepositoryImportRequest"},"RepositoryRead":{"properties":{"user_id":{"type":"integer","title":"User Id"},"owner":{"type":"string","title":"Owner"},"name":{"type":"string","title":"Name"},"github_url":{"type":"string","title":"Github Url"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"primary_language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Primary Language"},"stars":{"type":"integer","title":"Stars","default":0},"forks":{"type":"integer","title":"Forks","default":0},"default_branch":{"type":"string","title":"Default Branch","default":"main"},"health_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Health Score"},"id":{"type":"integer","title":"Id"},"imported_at":{"type":"string","format":"date-time","title":"Imported At"}},"type":"object","required":["user_id","owner","name","github_url","id","imported_at"],"title":"RepositoryRead"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","username","email","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}